requirements.txt
botocore==1.15.46
aiobotocore==1.0.4
aioredis==1.3.1
clamd==1.0.2
fabric==2.5.0
Flask==1.1.2
gunicorn==20.0.4
manhattan-assets==0.6.6
manhattan-chains==0.0.15
manhattan-comparable==0.0.15
manhattan-dispatch==0.0.6
manhattan-formatters==0.2.0
manhattan-forms==0.1.17
manhattan-mail==0.0.13
manhattan-manage==0.11.1
manhattan-nav==0.0.18
manhattan-publishing==0.2.2
manhattan-secure==0.1.2
manhattan-users==0.0.34
manhattan-utils==0.0.8
mutagen==1.44.0
MongoFrames==1.3.6
numpy==1.18.3
opencv-python-headless==4.2.0.34
pyotp==2.3.0
python-slugify==4.0.0
qrcode==6.1
redis==3.4.1
patchwork==1.0.1
Pillow==7.1.1
pip==20.2.4
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
sentry-sdk==0.14.3
shortuuid==1.0.1
swm==0.0.18
tornado==6.0.4
